% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_geotiff.R
\name{add_geotiff}
\alias{add_geotiff}
\title{Add a tif to a leaflet function (tweaked leafem::addGeotiff)}
\usage{
add_geotiff(
  map,
  file = NULL,
  url = NULL,
  group = NULL,
  layerId = NULL,
  resolution = 96,
  bands = NULL,
  arith = NULL,
  project = TRUE,
  method = NULL,
  opacity = 0.8,
  options = leaflet::tileOptions(),
  colorOptions = NULL,
  rgb = FALSE,
  pixelValuesToColorFn = NULL,
  autozoom = TRUE,
  imagequery = TRUE,
  imagequeryOptions = NULL,
  ...
)
}
\arguments{
\item{file}{path to the GeoTIFF file to render.}

\item{url}{url to the GeoTIFF file to render. Ignored if \code{file} is provided.}

\item{group}{he name of the group this raster image should belong to.}

\item{layerId}{the layerId.}

\item{resolution}{the target resolution for the simple nearest neighbor interpolation.
Larger values will result in more detailed rendering, but may impact performance.
Default is 96 (pixels).}

\item{bands}{which bands to use in case of multi-band Geotiff.}

\item{arith}{an optional function to be applied to a multi-layer object.
Will be computed on-the-fly in the browser.}

\item{project}{\strong{Change from leafem version:} default is FALSE and expects epsg:4326 rasters.}

\item{method}{character defining the resampling method to be used when
\code{project} is \code{TRUE}.
See \url{https://gdal.org/en/latest/programs/gdalwarp.html#cmdoption-gdalwarp-r} for
possible values.}

\item{opacity}{opacity of the rendered layer.}

\item{options}{options to be passed to the layer.
See \code{\link[leaflet]{tileOptions}} for details.}

\item{colorOptions}{list defining the palette, breaks and na.color to be used.}

\item{rgb}{logical, whether to render Geotiff as RGB.}

\item{pixelValuesToColorFn}{optional JS function to be passed to the browser.
Can be used to fine tune and manipulate the color mapping.
See examples & \url{https://github.com/r-spatial/leafem/issues/25} for
some examples.}

\item{autozoom}{whether to automatically zoom to the full extent of the layer.
Default is \code{TRUE}}

\item{imagequery}{If \code{TRUE} a leaflet control with the hovered/clicked
value will appear on the map.}

\item{imagequeryOptions}{additional options for the control panel.}

\item{...}{currently not used.}
}
\value{
A leaflet map object.
}
\description{
This version uses file.copy instead of sf::gdal_utils to create the temp .tif that is drawn, to avoid huge file sizes (see \href{https://github.com/dew-landscapes/envSDMs/issues/7}{envSDMs #7})
Designed to work with rasters in 4326 and therefore this version uses project = FALSE as default; leafem uses TRUE and reprojects on the fly.
}
\details{
@param map the map to add the raster data to.
}
