% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tif.R
\name{add_tif}
\alias{add_tif}
\title{Methods for adding a geotiff to a leaflet map}
\usage{
add_tif(map, use_reproject, reproj_file, original_file, prefix, ...)
}
\arguments{
\item{map}{Leaflet map, usually via \verb{leafletProxy() |> }}

\item{use_reproject}{Logical: one or more tests for if a COG should be used}

\item{reproj_file}{File path to the COG}

\item{original_file}{File path to the original tif, for when use_reproject == FALSE}

\item{prefix}{Prefix used in \code{shiny::addResourcePath}; forms part of the url in addCOG}

\item{...}{Passed to the leafem function, e.g. visual options. Note that some arguments (e.g. \code{colorOptions}) aren't currently supported by addCOG.}
}
\description{
Calls either \code{addCOG} or \code{addGeotiff} (from leafem) to add a tif file to a leaflet map, based on
some test (\code{use_reproject}) of if a reprojected COG exists.
}
\examples{
\dontrun{
 add_tif(use_reproject = loaded$predmaps == "fine" && input$select_tif_type == "reproject" && sel_taxa_files()$predict_reproj_exists,
         reproj_file = sel_taxa_files()$predict_reproj_file,
         original_file = sel_taxa_files()$predict_fine_file,
         group = "Prediction")
}



}
