% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLegendCustom.R
\name{addLegendCustom}
\alias{addLegendCustom}
\title{Add custom legend to leaflet map}
\usage{
addLegendCustom(
  map,
  colors,
  labels,
  sizes = 14,
  shapes = "square",
  borders = colors,
  opacity = 1,
  proxy = TRUE,
  ...
)
}
\arguments{
\item{map}{Leaflet map object, optionally via leafletProxy (see 'proxy')}

\item{colors}{Vector of colour names or hex codes. Fill colour(s)}

\item{labels}{Character vector. Legend label(s)}

\item{sizes}{Numeric vector of legend shape size. Default 14 most closely matches default 'unstyled' leaflet legend.}

\item{shapes}{Vector of shapes. Only supports "square" or "circle", as shape comes only from the border-radius style of squares.}

\item{borders}{Vector of colour names or hex codes. Shape outline colour(s)}

\item{opacity}{Numeric between 0-1, length 1. Opacity of legend shapes}

\item{proxy}{Use `leafletProxy(map)`? Use only in a shiny server.}

\item{...}{Other arguments to `leaflet::addLegend` (e.g. title)}
}
\value{
Leaflet legend added to the `map` provided
}
\description{
Default leaflet legends only exist as a colour-filled square, this function allows customisation of size, shape, fill, and outline.
If style parameters are of unequal lengths they get recycled.
From https://stackoverflow.com/questions/52812238/custom-legend-with-r-leaflet-circles-and-squares-in-same-plot-legends
}
