% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_basemap.R
\name{build_basemap}
\alias{build_basemap}
\title{Create a standard leaflet base map}
\usage{
build_basemap(
  panes = NULL,
  panes_zIndex = NULL,
  use_basemaps = c("OpenStreetMap", "ImageMapSA", "Esri imagery"),
  ...
)
}
\arguments{
\item{panes}{Character vector of names to add as additional height-ordered layers. Call in added layers with options=pathOptions(pane = 'blah'). Basemaps are automatically included at zIndex=1.}

\item{panes_zIndex}{Numeric vector of panes levels. Must be equal length to `panes`.}

\item{use_basemaps}{Names of leaflet basemaps. Defaults are OpenStreetMap & ESRI World Imagery (built in to leaflet), and ImageMapSA via its public URL (which reverts to ESRI imagery outside its extent).}

\item{...}{Passed to initial leaflet(), eg. for setting height.}
}
\value{
A leaflet map, as an HTML widget object.
}
\description{
Create a leaflet map upon which to add layers through `leaflet::add` functions or `leafletProxy`.
Includes controls for panes zIndex (i.e. layer heights), basemaps, and sets controls positions.
}
