% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_basemap_sidebyside.R
\name{build_basemap_sidebyside}
\alias{build_basemap_sidebyside}
\title{Create a standardised leaflet base map for two layers side-by-side}
\usage{
build_basemap_sidebyside(
  panes = NULL,
  panes_zIndex = NULL,
  basemap = c("OpenStreetMap", "ImageMapSA", "Esri imagery"),
  ...
)
}
\arguments{
\item{panes}{Character vector of names to add as additional height-ordered layers. Call in added layers with \code{options=pathOptions(pane = 'blah')}}

\item{panes_zIndex}{Numeric vector of panes levels. Must be equal length to \code{panes}.}

\item{basemap}{Name of basemap layer to use. Note that \code{sidebyside} doesn't support layerControl for basemaps, so only one value is used. Supports default OpenStreetMap & ESRI World Imagery, and ImageMapSA via its public URL (which reverts to ESRI imagery outside its extent).}

\item{...}{Passed to initial \code{leaflet()}, eg. for setting height.}
}
\value{
A leaflet map, as an HTML widget object.
}
\description{
Create a leaflet map upon which to add layers through \code{leaflet::add} functions or \code{leafletProxy}.
Uses \link[leaflet.extras2:sidebyside]{leaflet.extras2::sidebyside} to add a vertical slider to view two layers side-by-side.
Add additional layers to a single side by specifying \code{options = pathOptions(pane = 'left/right')}, or add layers above the slider by using the panes parameters with \code{zIndex} >1.
}
\details{
Includes fewer options than \link{build_basemap}, which is better for adding multiple layers.
}
