% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shp.R
\name{download_shp}
\alias{download_shp}
\title{Download a shapefile from a shiny app - superceded.}
\usage{
download_shp(
  data,
  layer_names,
  zip_name = if (!multiple) layer_names,
  multiple = FALSE
)
}
\arguments{
\item{data}{Data to be downloaded, usually a reactive. If multiple = TRUE, must be a list.}

\item{layer_names}{Name of the shapefile/s, ie will output layer_name.shp, layer_name.dbf, etc. If multiple = TRUE, should be same length as data list.}

\item{zip_name}{Name of the .zip folder downloaded.}

\item{multiple}{Download contains multiple shapefiles? Check requirements for data and layer_names if TRUE.}
}
\value{
A .zip file download from the shiny server, containing one or more shapefiles.
}
\description{
Due to bugs in reactive tracking, superceded by \code{envShiny::make_shp_download.R}
}
\details{
Create and zip the components of an ESRI shapefile (i.e. .shp, .dbf, etc) into a single download. Use instead of shiny::downloadHandler
From <https://stackoverflow.com/questions/41707760/download-a-shape-file-from-shiny>
}
