% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shp_multiple.R
\name{download_shp_multiple}
\alias{download_shp_multiple}
\title{Download several shapefiles from a shiny server}
\usage{
download_shp_multiple(data, layer_names = names(data), zip_name)
}
\arguments{
\item{data}{A list object of datasets to download. Must be named unless layer_names is specified.}

\item{layer_names}{Names to save individual layers as. Length must match length of data - usually should use names(data) or some derivitive, e.g `paste0("data_", names(data))``}

\item{zip_name}{Name of the .zip folder downloaded.}
}
\value{
A .zip download from the shiny server containing all supplied data.
}
\description{
Writes individual shapefiles (with all necessary components, i.e. .shp, .dbf etc), and zips them into a single download.
}
\seealso{
[download_shp()]
}
