% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_html_list.R
\name{format_html_list}
\alias{format_html_list}
\title{Format a character vector to a bulleted/numbered list}
\usage{
format_html_list(char, ordered = FALSE)
}
\arguments{
\item{char}{Character vector of items to be listed}

\item{ordered}{Use numbers (TRUE) or bullets (FALSE; default)}
}
\value{
HTML character string.
}
\description{
Format a character vector to a bulleted/numbered list
}
