% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_colourise.R
\name{html_colourise}
\alias{html_colourise}
\title{Colour text for use in HTML output (eg help docs)}
\usage{
html_colourise(text, colour)
}
\arguments{
\item{text}{Text to display}

\item{colour}{Valid css colour}
}
\value{
Character string
}
\description{
Colour text for use in HTML output (eg help docs)
}
\details{
From https://bookdown.org/yihui/rmarkdown-cookbook/font-color.html
}
