% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_help.R
\name{knit_help}
\alias{knit_help}
\title{Knit help files}
\usage{
knit_help(help_dir = here::here("shiny", "help"))
}
\arguments{
\item{help_dir}{Folder containing .rmd files.}
}
\value{
.md files, saved into help_dir
}
\description{
Knit .md files from .rmd files for display in shiny via includeMarkdown().
}
