% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_bib_tbl.R
\name{make_bib_tbl}
\alias{make_bib_tbl}
\title{Prepare a bio_all.bib file for inclusion in a metadata bibliography}
\usage{
make_bib_tbl(bibfile)
}
\arguments{
\item{bibfile}{Path to .bib file}
}
\value{
Dataframe with cols 'data_name', 'ref' (plain text) and 'bibtex'
}
\description{
Uses a single .bib file to gather data names and their respective bibtex and reference-list formats.
}
\examples{
\dontrun{
make_bib_tbl("H:/data/envOcc/aus_imcra_prov_dissolve______0__P50Y/bio_all.bib")
}

}
