% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newtab_link.R
\name{newtab_link}
\alias{newtab_link}
\title{Create a 'new tab' hyperlink}
\usage{
newtab_link(text, url, name = NULL)
}
\arguments{
\item{text}{Clickable text to appear on the page. Is followed by a new tab icon.}

\item{url}{Link to open.}

\item{name}{Text to show when hovering over link text.}
}
\value{
An HTML string that is displayed as a hyperlink
}
\description{
Takes a text string, appends an icon, and handles the html to open in a new tab.
}
