% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncate_string.R
\name{truncate_string}
\alias{truncate_string}
\title{Truncate long text strings in a DT::datatable output}
\usage{
truncate_string(cols = "_all", length = 60)
}
\arguments{
\item{cols}{Character vector of columns to apply truncation to}

\item{length}{Number of characters to truncate to}
}
\value{
A list, to be used within `columnDefs` list in datatable `options` (see example)
}
\description{
Truncate long text strings in a DT::datatable output
}
\examples{
\dontrun{
DT::datatable(
data.frame(text = "A very very very very very long string"),
options = list(
  columnDefs = list(
    truncate_string("text", 20)
  )
)
)
}

}
