% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sidebyside_basemap.R
\name{update_sidebyside_basemap}
\alias{update_sidebyside_basemap}
\title{Update the base layer used in a side-by-side leaflet map}
\usage{
update_sidebyside_basemap(
  map,
  basemap = c("OpenStreetMap", "ImageMapSA", "Esri imagery")
)
}
\arguments{
\item{map}{An existing leaflet map object, usually piped from \code{leafletProxy()}}

\item{basemap}{The basemap layer which the map should be updated to use}
}
\value{
A leaflet map, as an HTML widget object.
}
\description{
Used after creating a map using \link{build_basemap_sidebyside}. Needed because \link[leafem:addsidebyside]{leafem::addsidebyside} doesn't support the use of in-map base layers controls.
}
\details{
If another layer (e.g. a tif) is being drawn on top of the basemaps, and is 'swipeable', use
\code{map |> hideGroup("tif") |> showGroup("tif")} to ensure it stays on top of the basemaps.
}
