% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withNotification.R
\name{withNotification}
\alias{withNotification}
\title{Show a shiny notification while code runs}
\usage{
withNotification(expr, message, detail = NULL, ...)
}
\arguments{
\item{expr}{Code to execute}

\item{message}{Main message to display in notification panel. Shown in bold with strong().}

\item{detail}{Secondary message to display after message, in smaller text}

\item{...}{to showNotification (e.g. type)}
}
\description{
Uses the function format of shiny::withProgress (i.e. wrap an expression in this function), but without a progress bar.
Useful when a long-running function can't be broken down into smaller chunks (e.g. loading a large RDS or target).
}
