% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shp.R
\name{download_shp}
\alias{download_shp}
\title{Download a shapefile from a shiny app}
\usage{
download_shp(data, layer_name, zip_name = layer_name)
}
\arguments{
\item{data}{Data to be downloaded, usually a reactive}

\item{layer_name}{Name of the shapefile, ie will output layer_name.shp, layer_name.dbf, etc.}

\item{zip_name}{Name of the .zip folder downloaded.}
}
\value{
A file download from the shiny server.
}
\description{
Create and zip the components of an ESRI shapefile (i.e. .shp, .dbf, etc) into a single download. Use instead of shiny::downloadHandler
From https://stackoverflow.com/questions/41707760/download-a-shape-file-from-shiny
}
\seealso{
[download_shp_multiple()]
}
