% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aoi.R
\name{plot_aoi}
\alias{plot_aoi}
\title{Plot area of interest}
\usage{
plot_aoi(extent_sf, extent_poly = NULL, region_taxa_sf = NULL, zoom = FALSE)
}
\arguments{
\item{extent_sf}{An \code{sf} object of the actual extent i.e. the AOI, e.g. from \code{tar_read(extent_sf)}.}

\item{extent_poly}{An \code{sf} object of the base vector used to build \code{extent_sf}, e.g. from H:/data/vector.}

\item{region_taxa_sf}{An \code{sf} object to add to the map, e.g. from tar_read(region_taxa_sf)}

\item{zoom}{Logical. Set the map extent to \code{extent_sf}? Works well with a \code{bslib::input_switch} input.}
}
\value{
ggplot2 object with \code{extent_sf} overlayed on an SA/Aus map, with optional \code{extent_poly} and \code{region_taxa}.
}
\description{
Create a simple ggplot of an AOI and, optionally, the 'base' vector used to create the aoi without buffers/filtering, and a region_taxa polygon.
Map extent is set to the wider of SA or the AOI, or can be set manually to just the AOI with \code{zoom}.
}
\details{
Requires \code{sa} and \code{aus} sf objects to build the map - usually as found in data/vector. They are NOT included in this package. Map crs is set to \code{crs(aus)}.
}
